requirements.txt
aiohttp==3.7.4.post0
aioredis==2.0.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
h11==0.13.0
idna==3.3
multidict==6.0.2
pycparser==2.21
PyJWT==2.3.0
sniffio==1.2.0
starlette==0.18.0
typing-extensions==4.1.1
uvicorn==0.17.4
yarl==1.7.2
